LayoutTests/fast/canvas/webgl/angle-instanced-arrays-out-of-bounds-expected.txt

CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: no ELEMENT_ARRAY_BUFFER bound
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: no ELEMENT_ARRAY_BUFFER bound
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawArraysInstanced: first or count < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_VALUE: drawElementsInstanced: count or offset < 0
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: request out of bounds for current ELEMENT_ARRAY_BUFFER
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: no ELEMENT_ARRAY_BUFFER bound
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: no ELEMENT_ARRAY_BUFFER bound
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawArraysInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
CONSOLE MESSAGE: line 1: WebGL: INVALID_OPERATION: drawElementsInstanced: attempt to access out of bounds arrays
Test of drawArraysInstancedANGLE and drawElementsInstancedANGLE with out-of-bounds parameters

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

 PASS Successfully enabled ANGLE_instanced_arrays extension

Test with 1 instance without instanced attributes

Test empty buffer
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 1, 1) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 10000, 1) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 10000000000000, 1) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 1, 0, 1) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 0, 1) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 100, 0, 1) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, -1, 1) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, -1, 0, 1) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 1, -1, 1) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, -1, 1, 1) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 0xffffffff, 1) was expected value: INVALID_VALUE.

Test buffer with 3 float vectors
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 3, 1) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 3, 2, 1) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 10000, 1) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 10000000000000, 1) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 0, 1) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 100, 0, 1) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, -1, 1) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, -1, 0, 1) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 1, -1, 1) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, -1, 1, 1) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 0xffffffff, 1) was expected value: INVALID_VALUE.

Test buffer with interleaved (3+2) float vectors
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 9, 1) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, -500, 1) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, -200, 1, 1) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, -200, -500, 1) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 0xffffffff, 1) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0xffffffff, 0xffffffff, 1) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0xffffffff, 1, 1) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 200, 1) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 0x7fffffff, 1) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0x7fffffff, 1, 1) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0x7fffffff, 0x7fffffff, 1) was expected value: INVALID_OPERATION.

Test empty index buffer
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0, gl.UNSIGNED_BYTE, 0, 1) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 3, gl.UNSIGNED_BYTE, 0, 1) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 10000, gl.UNSIGNED_BYTE, 0, 1) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 10000000000000, gl.UNSIGNED_BYTE, 0, 1) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 1, gl.UNSIGNED_BYTE, 0, 1) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, -1, gl.UNSIGNED_BYTE, 0, 1) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0, gl.UNSIGNED_BYTE, -1, 1) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, -1, gl.UNSIGNED_BYTE, 1, 1) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 1, gl.UNSIGNED_BYTE, -1, 1) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0xffffffff, gl.UNSIGNED_BYTE, 0, 1) was expected value: INVALID_VALUE.

Test buffer with 3 byte indexes
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 3, gl.UNSIGNED_BYTE, 0, 1) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 3, gl.UNSIGNED_BYTE, 2, 1) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 10000, gl.UNSIGNED_BYTE, 0, 1) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 10000000000000, gl.UNSIGNED_BYTE, 0, 1) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, -1, gl.UNSIGNED_BYTE, 0, 1) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0, gl.UNSIGNED_BYTE, -1, 1) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, -1, gl.UNSIGNED_BYTE, 1, 1) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 1, gl.UNSIGNED_BYTE, -1, 1) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0xffffffff, gl.UNSIGNED_BYTE, 0, 1) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0, gl.UNSIGNED_BYTE, 4, 1) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0x7fffffff, gl.UNSIGNED_BYTE, 0, 1) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0x7fffffff, gl.UNSIGNED_BYTE, 0x7fffffff, 1) was expected value: INVALID_OPERATION.
PASS gl.bufferData(gl.ELEMENT_ARRAY_BUFFER, (new Uint8Array([ 3, 0, 1, 2 ])).subarray(1), gl.STATIC_DRAW) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 3, gl.UNSIGNED_BYTE, 0, 1) was expected value: NO_ERROR.
PASS gl.bufferSubData(gl.ELEMENT_ARRAY_BUFFER, 0, new Uint8Array([ 3, 0, 1])) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 3, gl.UNSIGNED_BYTE, 0, 1) was expected value: INVALID_OPERATION.
PASS gl.bufferSubData(gl.ELEMENT_ARRAY_BUFFER, 0, (new Uint8Array([ 3, 0, 1, 2 ])).subarray(1)) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 3, gl.UNSIGNED_BYTE, 0, 1) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0, gl.UNSIGNED_BYTE, 0, 1) was expected value: NO_ERROR.

Test buffer with interleaved (3+2) float vectors
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_SHORT, 0, 1) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_SHORT, 1000, 1) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 12, gl.UNSIGNED_SHORT, 0, 1) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 15, gl.UNSIGNED_SHORT, 0, 1) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 18, gl.UNSIGNED_SHORT, 0, 1) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 3, gl.UNSIGNED_SHORT, 30, 1) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0xffffffff, gl.UNSIGNED_SHORT, 0, 1) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 1, gl.UNSIGNED_SHORT, 0xffffffff, 1) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0x7fffffff, gl.UNSIGNED_SHORT, 0, 1) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0, gl.UNSIGNED_SHORT, 0, 1) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 6, gl.UNSIGNED_SHORT, 0, 1) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 6, gl.UNSIGNED_SHORT, 1, 1) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 6, gl.UNSIGNED_SHORT, 2, 1) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 6, gl.UNSIGNED_SHORT, 0, 1) was expected value: INVALID_OPERATION.

Test buffer setting attrib 0 to a buffer too small and disable it.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 6, gl.UNSIGNED_SHORT, 2, 1) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 6, gl.UNSIGNED_SHORT, 2, 1) was expected value: INVALID_OPERATION.

Test with 2 instances without instanced attributes

Test empty buffer
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 1, 2) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 10000, 2) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 10000000000000, 2) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 1, 0, 2) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 0, 2) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 100, 0, 2) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, -1, 2) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, -1, 0, 2) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 1, -1, 2) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, -1, 1, 2) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 0xffffffff, 2) was expected value: INVALID_VALUE.

Test buffer with 3 float vectors
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 3, 2) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 3, 2, 2) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 10000, 2) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 10000000000000, 2) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 0, 2) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 100, 0, 2) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, -1, 2) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, -1, 0, 2) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 1, -1, 2) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, -1, 1, 2) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 0xffffffff, 2) was expected value: INVALID_VALUE.

Test buffer with interleaved (3+2) float vectors
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 9, 2) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, -500, 2) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, -200, 1, 2) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, -200, -500, 2) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 0xffffffff, 2) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0xffffffff, 0xffffffff, 2) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0xffffffff, 1, 2) was expected value: INVALID_VALUE.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 200, 2) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 0x7fffffff, 2) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0x7fffffff, 1, 2) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0x7fffffff, 0x7fffffff, 2) was expected value: INVALID_OPERATION.

Test empty index buffer
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0, gl.UNSIGNED_BYTE, 0, 2) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 3, gl.UNSIGNED_BYTE, 0, 2) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 10000, gl.UNSIGNED_BYTE, 0, 2) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 10000000000000, gl.UNSIGNED_BYTE, 0, 2) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 1, gl.UNSIGNED_BYTE, 0, 2) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, -1, gl.UNSIGNED_BYTE, 0, 2) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0, gl.UNSIGNED_BYTE, -1, 2) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, -1, gl.UNSIGNED_BYTE, 1, 2) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 1, gl.UNSIGNED_BYTE, -1, 2) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0xffffffff, gl.UNSIGNED_BYTE, 0, 2) was expected value: INVALID_VALUE.

Test buffer with 3 byte indexes
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 3, gl.UNSIGNED_BYTE, 0, 2) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 3, gl.UNSIGNED_BYTE, 2, 2) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 10000, gl.UNSIGNED_BYTE, 0, 2) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 10000000000000, gl.UNSIGNED_BYTE, 0, 2) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, -1, gl.UNSIGNED_BYTE, 0, 2) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0, gl.UNSIGNED_BYTE, -1, 2) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, -1, gl.UNSIGNED_BYTE, 1, 2) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 1, gl.UNSIGNED_BYTE, -1, 2) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0xffffffff, gl.UNSIGNED_BYTE, 0, 2) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0, gl.UNSIGNED_BYTE, 4, 2) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0x7fffffff, gl.UNSIGNED_BYTE, 0, 2) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0x7fffffff, gl.UNSIGNED_BYTE, 0x7fffffff, 2) was expected value: INVALID_OPERATION.
PASS gl.bufferData(gl.ELEMENT_ARRAY_BUFFER, (new Uint8Array([ 3, 0, 1, 2 ])).subarray(1), gl.STATIC_DRAW) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 3, gl.UNSIGNED_BYTE, 0, 2) was expected value: NO_ERROR.
PASS gl.bufferSubData(gl.ELEMENT_ARRAY_BUFFER, 0, new Uint8Array([ 3, 0, 1])) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 3, gl.UNSIGNED_BYTE, 0, 2) was expected value: INVALID_OPERATION.
PASS gl.bufferSubData(gl.ELEMENT_ARRAY_BUFFER, 0, (new Uint8Array([ 3, 0, 1, 2 ])).subarray(1)) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 3, gl.UNSIGNED_BYTE, 0, 2) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0, gl.UNSIGNED_BYTE, 0, 2) was expected value: NO_ERROR.

Test buffer with interleaved (3+2) float vectors
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_SHORT, 0, 2) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_SHORT, 1000, 2) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 12, gl.UNSIGNED_SHORT, 0, 2) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 15, gl.UNSIGNED_SHORT, 0, 2) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 18, gl.UNSIGNED_SHORT, 0, 2) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 3, gl.UNSIGNED_SHORT, 30, 2) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0xffffffff, gl.UNSIGNED_SHORT, 0, 2) was expected value: INVALID_VALUE.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 1, gl.UNSIGNED_SHORT, 0xffffffff, 2) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0x7fffffff, gl.UNSIGNED_SHORT, 0, 2) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 0, gl.UNSIGNED_SHORT, 0, 2) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 6, gl.UNSIGNED_SHORT, 0, 2) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 6, gl.UNSIGNED_SHORT, 1, 2) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 6, gl.UNSIGNED_SHORT, 2, 2) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 6, gl.UNSIGNED_SHORT, 0, 2) was expected value: INVALID_OPERATION.

Test buffer setting attrib 0 to a buffer too small and disable it.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 6, gl.UNSIGNED_SHORT, 2, 2) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 6, gl.UNSIGNED_SHORT, 2, 2) was expected value: INVALID_OPERATION.

Test out-of-range instanced attributes

Test with an empty buffer for the instanced attribute
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 9, 0) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 10000, 0) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 9, 1) was expected value: INVALID_OPERATION.
Test with a buffer with 1 float for the instanced attribute
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 9, 0) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 10000, 0) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 9, 1) was expected value: INVALID_OPERATION.

Test with a buffer with 2 floats for the instanced attribute
Divisor 1
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 9, 1) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 9, 2) was expected value: INVALID_OPERATION.
Divisor 3
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 9, 3) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 9, 4) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 9, 10000) was expected value: INVALID_OPERATION.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 9, 0x7fffffff) was expected value: INVALID_OPERATION.

Test with a buffer with 4 floats for the instanced attribute
Divisor 1
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 9, 2) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 9, 3) was expected value: INVALID_OPERATION.
Divisor 2
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 9, 4) was expected value: NO_ERROR.
PASS ext.drawArraysInstancedANGLE(gl.TRIANGLES, 0, 9, 5) was expected value: INVALID_OPERATION.

Test out-of-range instanced attributes

Test with an empty buffer for the instanced attribute
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_BYTE, 0, 0) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 10000, gl.UNSIGNED_BYTE, 0, 0) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_BYTE, 0, 1) was expected value: INVALID_OPERATION.
Test with a buffer with 1 float for the instanced attribute
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_BYTE, 0, 0) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 10000, gl.UNSIGNED_BYTE, 0, 0) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_BYTE, 0, 1) was expected value: INVALID_OPERATION.

Test with a buffer with 2 floats for the instanced attribute
Divisor 1
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_BYTE, 0, 1) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_BYTE, 0, 2) was expected value: INVALID_OPERATION.
Divisor 3
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_BYTE, 0, 3) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_BYTE, 0, 4) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_BYTE, 0, 10000) was expected value: INVALID_OPERATION.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_BYTE, 0, 0x7fffffff) was expected value: INVALID_OPERATION.

Test with a buffer with 4 floats for the instanced attribute
Divisor 1
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_BYTE, 0, 2) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_BYTE, 0, 3) was expected value: INVALID_OPERATION.
Divisor 2
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_BYTE, 0, 4) was expected value: NO_ERROR.
PASS ext.drawElementsInstancedANGLE(gl.TRIANGLES, 9, gl.UNSIGNED_BYTE, 0, 5) was expected value: INVALID_OPERATION.

PASS successfullyParsed is true

TEST COMPLETE
